# api/openapi/prior-auth-api.yaml
openapi: 3.1.0
info:
  title: Cloud Health Office — Prior Authorization API
  version: 4.0.0
  description: |
    FHIR R4 Prior Authorization API with CDS Hooks for CMS-0057-F compliance.

    Implements electronic prior authorization aligned with the Da Vinci Prior
    Authorization Support (PAS) Implementation Guide and Coverage Requirements
    Discovery (CRD) / Documentation Templates and Rules (DTR) for PARDD.

    **CMS-0057-F requires** payers to support electronic prior auth submission,
    status inquiry, and real-time coverage requirements discovery by January 1, 2027.

    **Key Capabilities:**
    - FHIR Claim-based prior auth submission (replaces X12 278)
    - Real-time auth status inquiry via ClaimResponse
    - CDS Hooks for coverage requirements discovery (CRD)
    - Bidirectional X12 278 ↔ FHIR mapping
    - SLA tracking (urgent: 24h, standard: 7 days per CMS timelines)
    - Clinical attachment support (DocumentReference, Binary)
    - Clearinghouse packaging for hybrid X12/FHIR workflows
    - Decision timeline audit trail

    **References:**
    - CMS-0057-F Final Rule (Prior Authorization Requirements)
    - Da Vinci PAS (Prior Authorization Support) IG
    - Da Vinci CRD (Coverage Requirements Discovery) IG
    - Da Vinci DTR (Documentation Templates and Rules) IG
    - CDS Hooks Specification v2.0
    - X12 278 Health Care Services Review
  contact:
    name: Aurelianware, Inc.
    url: https://cloudhealthoffice.com
  license:
    name: Apache 2.0
...
servers:
...
security:
  - smart_on_fhir: [user/Claim.write, user/ClaimResponse.read]
...
paths:
...
  /ClaimResponse:
    get:
      operationId: searchAuthorizations
      summary: Search prior authorizations
      description: |
        Search for prior authorizations by patient, provider, status, or date.
        CMS-0057-F requires payers to make authorization status available
        to both providers and patients.
      tags: [Prior Authorization]
      parameters:
        - name: patient
          in: query
          schema:
            type: string
          example: "Patient/pat-12345"
        - name: provider
          in: query
          schema:
            type: string
        - name: outcome
          in: query
          schema:
            type: string
            enum: [complete, queued, partial, error]
        - name: created
          in: query
          schema:
            type: string
          description: Date range filter
        - name: _count
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Bundle of ClaimResponse resources
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Bundle'

  # ── CDS Hooks — Coverage Requirements Discovery ──────────
...
components:
  schemas:
...
    Bundle:
      type: object
      properties:
        resourceType:
          type: string
          const: Bundle
        type:
          type: string
        total:
          type: integer
        entry:
          type: array
...
  securitySchemes:
    smart_on_fhir:
      type: oauth2
      description: SMART on FHIR for provider EHR integration
      flows:
        authorizationCode:
          authorizationUrl: https://login.cloudhealthoffice.com/oauth2/authorize
          tokenUrl: https://login.cloudhealthoffice.com/oauth2/token
          scopes:
            user/Claim.write: Submit prior authorization requests
            user/Claim.read: Read authorization requests
            user/ClaimResponse.read: Read authorization decisions
            user/DocumentReference.write: Submit clinical attachments
            openid: OpenID Connect
            fhirUser: FHIR user identity